requirements/dev.txt
django-extensions==3.2.1
ipython==8.5.0
faker==15.1.1
